drm: probe the GBM device in all cases
If we use the stock Mesa, the tegra display device will be capable of
performing rendering due to the presence of the software rendering
module. Assume the display device is not capable of rendering in all
cases, so that the Nouveau device gets a chance to be probed in that
case.